feat: Replace placeholder game cycle methods (WIP)
Placeholder methods like playTurn() and start() were designed
while thinking of each game phase as an independent loop.
This is not a good approach because of conflicts with QT's
render and event listening loops.
Therefore, the work has been started on replacing those
methods with slot/signal system that doesn't interfere with QT.