refactor: Use m_card field in EffectActivator
This allows easier binding of an EffectActivator to the specific
card that was passed as an argument to EffectActivator's
constructor. Since the Card class is abstract, we use a pointer.
Also, in order to forbid the initialization of an EffectActivator
object with no arguments (which would make m_card == nullptr),
we delete the default constructor.
This way, an EffectActivator object must have the card that its
responsible for.