feat: Add deserialization support in Game
We made our own convention for serialization and deserialization,
where, before sending any data, we need to first send a header
string that will somehow be an indentification of the data
that will come after it. That way allows us to map those header
strings into class methods that know how to deserialize data that
comes after them.
Currently, separate server program is used only for passing the
data from one client to the other(s).