fix: Segfault when clicking red spell/trap zones
The problem was that those zones were instantly colored red when
they were added to the scene. There is no need for that, it was
probably done because of some testing before. However, it allowed
for a segfault to happen because client's click on any of them
triggered onRedZoneClick which uses a global pointer to a card
that is about to be placed on field. However, that pointer only
gets set in onSummonButtonClick, which never happened. Because
of that, the pointer was null and therefore calling getCardType
produced a segfault.