refactor: Determine who plays first on the server
We have to move the function that was determining who plays first
to the server, which will calculate that using the random generator
and then send it to both players so that they have that information
in sync.
By doing this, we don't have the problem of both players calling
a separate random generator which often lead to a situation where
they had a different first player.