fix: Reposition button gets clicked twice
For some weird reason, even though all connects are done in the
same manner in onCardSelect, only Reposition button displays
this weird behaviour. Whenever its clicked, Qt acts like it was
double-clicked.
Using Qt::UniqueConnection as the 5th argument of connect() is
impossible since we are using lambda-slot (there is even a
warning about that). Therefore, this "dirty" way of using
card-local repositionCounter is used for now.