feat: Rework battle events, fix button connects
Battle events now use DESTROY_CARD header that can be used
for effects too, so this increases the reusability of the code.
Now, about the fix, the problem with double clicking of buttons
was that the connect was called in onCardSelect, which happened
every single time when a card was clicked. Now that code was
moved in onCardAddedToScene, which only happens once when
card gets added to the scene and ensures that thos card menu
connect only happen once, as they should.