fix: Various spell card activation bugs
There was a problem with spell cards not going
to the graveyard for the player that activated
them.
Also, in some specific cases there was a segfault,
for example when we used Monster Reborn to special
summon a monster from the graveyard. Before, we
only removed the monster that is to-be-summoned from
the hand, but this was a special case where we can
summon from the graveyard, so we needed to remove
it from the graveyard instead of the hand.
Ookazi and Dian Keto were also fixed, there was probably
an infinite notifications about finished deserialization
happening, since those cards have additional headers
sent to the server (specifically, LP_CHANGE and
EFFECT_ACTIVATED).
Finally, there was some missing code that was on develop
branch, but wasn't on main branch and that was just copy
pasted to main since git merges are misbehaving (when
doing a git merge, git says that everything is
"Already up to date", even though git diff shows
the changes), so a manual copy-paste "merge" was
deemed necessary.