fix: Fix SIGSEGV in the main menu
We had a problem with the usage of delete.
The m_pGame is initialized when the Start button is clicked,
and m_pGameSettings is initialized when the Game Settings
button is clicked.
However, the deletes for both of those were in mainmenu's
destructor, which gets called if we click the Quit button,
even if we didn't start the actual game or go into the settings.
Therefore, this causes the program to try to delete something
that hasn't had it's new() invoked yet.
We fix this simply by setting the default values for those ptrs
to null and then checking that in the destructor.